+++ GHEtool/gui/gui_structure.py
@@ -15,6 +15,7 @@ import PySide6.QtCore as QtC
 import pandas as pd
 
 from GHEtool import FOLDER
+from GHEtool.VariableClasses import FluidData, PipeData, Borehole
 from GHEtool.gui.gui_classes.translation_class import Translations
 from numpy import array, cos, int64, round, sin, sum
 from pandas import DataFrame as pd_DataFrame
@@ -1414,6 +1415,7 @@ class GUI(GuiStructure):
         self.option_pipe_number.widget.blockSignals(True)
         self.option_pipe_number.widget.setMaximum(min(max_nb_of_pipes, self.option_pipe_number.maximal_value))
         self.option_pipe_number.widget.blockSignals(False)
+        self.update_borehole_thermal_resistance()
 
     def update_borehole(self) -> None:
         """
@@ -1720,6 +1722,23 @@ class GUI(GuiStructure):
             data = f.readlines()
         data = [[float(val.replace(dec, ".")) for val in line.split(sep)] for line in data[1:]]
         self.custom_borefield.set_value(data)
+
+    def update_borehole_thermal_resistance(self) -> None:
+        """
+        This function updates the borehole thermal resistance.
+
+        Returns
+        -------
+        None
+        """
+        fluid_data = FluidData(self.option_fluid_mass_flow.get_value(), self.option_fluid_conductivity.get_value(),
+                               self.option_fluid_density.get_value(), self.option_fluid_capacity.get_value(), self.option_fluid_viscosity.get_value())
+        pipe_data = PipeData(self.option_pipe_grout_conductivity.get_value(), self.option_pipe_inner_radius.get_value(), self.option_pipe_outer_radius.get_value(),
+                             self.option_pipe_conductivity.get_value(), self.option_pipe_distance.get_value(), self.option_pipe_number.get_value(), self.option_pipe_roughness.get_value())
+
+        borehole = Borehole(fluid_data, pipe_data)
+        logging.info(f'The borehole thermal resistance is'
+                     f'{borehole.get_Rb(150, self.option_pipe_depth.get_value(), self.option_pipe_borehole_radius.get_value(), self.option_conductivity.get_value())}')
 
 
 def show_option_on_multiple_aims(first_aims: list[Aim], second_aims: list[Aim], option: Option):